+++ src/deltachat_cursed/main.py
@@ -185,9 +185,9 @@ def main() -> None:
         ac.add_account_plugin(log)
 
     if not ac.is_configured():
-        assert (
-            args.email
-        ), "you must specify --email once to configure this database/account"
+        if not args.email:
+            print("Error: You must specify --email once to configure the account")
+            sys.exit(1)
         ac.set_config("addr", args.email)
 
         if not args.password and is_oauth2(ac, args.email):
@@ -200,9 +200,9 @@ def main() -> None:
             flags |= deltachat.const.DC_LP_AUTH_OAUTH2  # noqa
             ac.set_config("server_flags", str(flags))
         else:
-            assert (
-                args.password
-            ), "you must specify --password once to configure this database/account"
+            if not args.password:
+                print("Error: You must specify --password once to configure the account")
+                sys.exit(1)
             ac.set_config("mail_pw", args.password)
 
         with ac.temp_plugin(ConfigureTracker(ac)) as tracker: